Revert "Fix #223. setTotal should take long as a parameter."
This reverts commit 62ea11b0c684f8ee715e5e31a4195a1c28b47a2a.

+++ src/com/esotericsoftware/kryo/io/Input.java
@@ -96,7 +96,7 @@ public class Input extends InputStream {
 	}
 
 	/** Sets the number of bytes read. */
-	public void setTotal (long total) {
+	public void setTotal (int total) {
 		this.total = total;
 	}
